requirements.txt
python-telegram-bot
tenacity
beautifulsoup4 >= 4.11.2
lxml